+++ plugins/function-tools/src/index.ts
@@ -1,3 +1,4 @@
 export { plugin, createFunctionToolsPlugin } from './plugin.js';
 export { buildAsyncFn, runFunction, type CompiledFn } from './compile.js';
+export { parsePackage, buildPackageFn, exportFn, PACKAGE_SEPARATOR, type ParsedPackage, type PackageExport } from './package.js';
 export { parseSignature, paramsSchema, type ParsedParam, type ParsedSignature } from './signature.js';